requirements.txt
chainlit==0.7.700
pandas==2.1.0
plotly==5.17.0
openai==1.3.0
requests==2.31.0
Flask==2.3.3
python-dotenv==1.0.0
numpy==1.24.3
logging==0.5.1.2
typing-extensions==4.7.1